moved logic from ImperativeViewController into new class ImperativeGestureReactor
